requirements.txt
streamlit
keras==2.3.1
tensorflow==2.2.0
opencv-python
pandas
facenet-pytorch
torch
torchvision
pillow
